chore(wmg): refactor BottomBanner 5948 (#5949)
