+++ frontend/src/components/BottomBanner/components/ModalContent/types.ts
@@ -1,0 +1,8 @@
+export interface Props {
+  id?: string;
+  isHubSpotReady?: boolean;
+  setError: (error: string) => void;
+  setEmail: (email: string) => void;
+  email: string;
+  emailValidationError?: string;
+}
