+++ frontend/src/components/BottomBanner/constants.ts
@@ -1,0 +1,22 @@
+export const FORM_CONTAINER_ID = "hubspot-form-container";
+export const FORM_CONTAINER_ID_QUERY = `#${FORM_CONTAINER_ID}`;
+export const FAILED_EMAIL_VALIDATION_STRING =
+  "Please provide a valid email address.";
+export const HIDDEN_NEWSLETTER_SUCCESS_MESSAGE =
+  "Thank you for joining our newsletter.";
+export const BOTTOM_BANNER_EXPIRATION_TIME_MS = 30 * 24 * 60 * 60 * 1000; // 30 days
+export const BOTTOM_BANNER_LAST_CLOSED_TIME_KEY = "bottomBannerLastClosedTime";
+export const NEWSLETTER_SIGNUP_TITLE = "Join Our Newsletter";
+export const NEWSLETTER_SIGNUP_MESSAGE = "Thanks for subscribing!";
+export const NEWSLETTER_SIGNUP_SUCCESS_MESSAGE =
+  "Get a quarterly email with the latest CELLxGENE features and data.";
+export const NEWSLETTER_SIGNUP_UNSUBSCRIBE_MESSAGE_UPON_SIGNUP =
+  'To unsubscribe, click on the "Unsubscribe" link at the bottom of the newsletter.';
+export const NEWSLETTER_SIGNUP_UNSUBSCRIBE_MESSAGE = "Unsubscribe at any time.";
+export const HUBSPOT_URL = "https://js.hsforms.net/forms/v2.js";
+export const NEWSLETTER_SIGNUP_BANNER_SUBSCRIBE_BUTTON_TEXT = "Subscribe";
+export const NEWSLETTER_SIGNUP_BANNER_SUBSCRIBE_TEXT =
+  "to our newsletter to receive updates about new features.";
+export const NEWSLETTER_SIGNUP_BANNER_SURVEY_LINK_TEXT = "quick survey.";
+export const NEWSLETTER_SIGNUP_BANNER_SURVEY_TEXT =
+  "Send us feedback with this";
